# api-docs.json
{"openapi":"3.0.1","servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"ApiKeyAuth":[]}],"tags":[{"name":"Reports","description":"Endpoints for generating session reports"},{"name":"Captures","description":"Endpoints for managing fingerprint captures"},{"name":"Sessions","description":"Endpoints to manage capture sessions"}],"paths":{"/sessions":{"post":{"tags":["Sessions"],"summary":"Create a new session","description":"Creates a contactless capture session for a user and device.","operationId":"createSession","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionCreateRequestDTO"}}},"required":true},"responses":{"400":{"description":"Invalid request payload","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionDTO"}}}},"200":{"description":"Session created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionDTO"}}}}}}},"/captures":{"post":{"tags":["Captures"],"summary":"Upload a capture image","description":"Uploads a fingerprint capture and calculates blur/brightness","operationId":"uploadCapture","parameters":[{"name":"sessionId","in":"query","description":"UUID of the session","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["image"],"type":"object","properties":{"image":{"type":"string","description":"Image file for the fingerprint capture","format":"binary"}}}}}},"responses":{"500":{"description":"Error processing capture","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid sessionId format","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Capture successfully uploaded","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/sessions/{sessionId}":{"get":{"tags":["Sessions"],"summary":"Get session by ID","description":"Fetch details of a session by its UUID.","operationId":"getSession","parameters":[{"name":"sessionId","in":"path","description":"UUID of the session","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Session retrieved successfully","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid sessionId format","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Session not found","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/reports/session/{sessionId}":{"get":{"tags":["Reports"],"summary":"Get report for a session","description":"Fetches a detailed report for a specific session, including capture statistics and metadata.","operationId":"getSessionReport","parameters":[{"name":"sessionId","in":"path","description":"UUID of the session","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Invalid sessionId format","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Session not found","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Report retrieved successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/captures/{captureId}":{"get":{"tags":["Captures"],"summary":"Fetch a capture by ID","description":"Retrieve a single capture using its unique captureId","operationId":"getCaptureById","parameters":[{"name":"captureId","in":"path","description":"UUID of the capture","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Capture not found","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid captureId format","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Capture retrieved successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/captures/user/{userId}":{"get":{"tags":["Captures"],"summary":"Fetch captures by userId","description":"Retrieve all captures for a specific user","operationId":"getCapturesByUserId","parameters":[{"name":"userId","in":"path","description":"UUID of the user","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"No captures found for the user","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid userId format","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Captures retrieved successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/captures/session/{sessionId}":{"get":{"tags":["Captures"],"summary":"Fetch captures by sessionId","description":"Retrieve all captures for a specific session","operationId":"getCapturesBySessionId","parameters":[{"name":"sessionId","in":"path","description":"UUID of the session","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Invalid sessionId format","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Captures retrieved successfully","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"No captures found for the session","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/captures/device/{deviceId}":{"get":{"tags":["Captures"],"summary":"Fetch captures by deviceId","description":"Retrieve all captures for a specific device","operationId":"getCapturesByDeviceId","parameters":[{"name":"deviceId","in":"path","description":"Device ID of the capturing device","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"No captures found for the device","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Captures retrieved successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"SessionCreateRequestDTO":{"required":["deviceId","userId"],"type":"object","properties":{"userId":{"type":"string","description":"User ID of the session owner (compact UUID as string)","example":"2a200aeccde847ca831cce8c40728701"},"deviceId":{"type":"string","description":"Device ID from which the session is initiated","example":"DEVICE123"}},"description":"Session creation request payload"},"CaptureDTO":{"type":"object","properties":{"captureId":{"type":"string","description":"Unique ID of the capture","format":"uuid","example":"1154e493-9ec5-4ffb-80c9-0f42b16949cc"},"sessionId":{"type":"string","description":"UUID of the session","format":"uuid","example":"2a200aec-cde8-47ca-831c-ce8c40728701"},"fingerPosition":{"type":"string","description":"Finger position of the capture","example":"THUMB","enum":["THUMB","INDEX","MIDDLE","RING","LITTLE"]},"blurScore":{"type":"number","description":"Blur score of the captured image","format":"float","example":120.5},"brightnessLevel":{"type":"string","description":"Brightness level of the captured image","example":"NORMAL"},"focusLocked":{"type":"boolean","description":"Indicates whether focus is locked","example":true},"livenessDetected":{"type":"boolean","description":"Indicates whether liveness was detected","example":true},"imageUrl":{"type":"string","description":"Public URL of the uploaded image","example":"http://localhost:8080/uploads/image.jpg"},"capturedAt":{"type":"string","description":"Timestamp when the capture was taken","format":"date-time","example":"2025-08-26T12:34:56.789Z"}},"description":"Data Transfer Object representing a fingerprint capture"},"SessionDTO":{"type":"object","properties":{"sessionId":{"type":"string","description":"Unique identifier of the session","format":"uuid","example":"2a200aec-cde8-47ca-831c-ce8c40728701"},"userId":{"type":"string","description":"User ID associated with the session","format":"uuid","example":"b1f4c3d5-e6a1-4b2f-9d7e-3f0a9a8f1234"},"userName":{"type":"string","description":"Name of the user","example":"John Doe"},"userEmail":{"type":"string","description":"Email of the user","example":"john.doe@example.com"},"deviceId":{"type":"string","description":"Device ID from which the session was created","example":"DEVICE123"},"createdAt":{"type":"string","description":"Session creation timestamp","example":"2025-08-26T18:14:23Z"},"captures":{"type":"array","description":"List of captures associated with this session","items":{"$ref":"#/components/schemas/CaptureDTO"}}},"description":"Data Transfer Object representing a session and its captures"}},"securitySchemes":{"ApiKeyAuth":{"type":"apiKey","name":"X-API-KEY","in":"header"}}}}